refactor: move request stores dtos
