+++ src/main/kotlin/com/snapshoes/store/presentation/dtos/request/store/UpdateStoreAddressDto.kt
@@ -1,4 +1,4 @@
-package com.snapshoes.store.presentation.dtos.request
+package com.snapshoes.store.presentation.dtos.request.store
 import com.snapshoes.store.presentation.dtos.response.common.AddressDto
 
 data class UpdateStoreAddressDto(
